fix: aws credentials_profile_name
