+++ apps/setting/models_provider/impl/aws_bedrock_model_provider/credential/embedding.py
@@ -11,21 +11,6 @@ from setting.models_provider.impl.aws_bedrock_model_provider.model.embedding imp
 
 class BedrockEmbeddingCredential(BaseForm, BaseModelCredential):
 
-    @staticmethod
-    def _update_aws_credentials(profile_name, access_key_id, secret_access_key):
-        credentials_path = os.path.join(os.path.expanduser("~"), ".aws", "credentials")
-        os.makedirs(os.path.dirname(credentials_path), exist_ok=True)
-
-        content = open(credentials_path, 'r').read() if os.path.exists(credentials_path) else ''
-        pattern = rf'\n*\[{profile_name}\]\n*(aws_access_key_id = .*)\n*(aws_secret_access_key = .*)\n*'
-        content = re.sub(pattern, '', content, flags=re.DOTALL)
-
-        if not re.search(rf'\[{profile_name}\]', content):
-            content += f"\n[{profile_name}]\naws_access_key_id = {access_key_id}\naws_secret_access_key = {secret_access_key}\n"
-
-        with open(credentials_path, 'w') as file:
-            file.write(content)
-
     def is_valid(self, model_type: str, model_name, model_credential: Dict[str, object], model_params, provider,
                  raise_exception=False):
         model_type_list = provider.get_model_type_list()
@@ -41,9 +26,6 @@ class BedrockEmbeddingCredential(BaseForm, BaseModelCredential):
             return False
 
         try:
-            self._update_aws_credentials('aws-profile', model_credential['access_key_id'],
-                                         model_credential['secret_access_key'])
-            model_credential['credentials_profile_name'] = 'aws-profile'
             model: BedrockEmbeddingModel = provider.get_model(model_type, model_name, model_credential)
             aa = model.embed_query('你好')
             print(aa)
